# backend/api-first/openapi.yaml
openapi: 3.1.0
info:
  title: KOSMO API
  version: "0.1.0"
  description: >-
    API de KOSMO - plataforma low-code de spec-driven development asistida por IA.
    Ciclo completo: Descubrimiento -> Caracteristicas -> Requisitos -> Modelo -> Codigo.

    Convenciones:
    - Autenticacion PKCE (RFC 7636) con JWT RS256 (bearerAuth).
    - Errores RFC 7807 (application/problem+json) con type URN, trace_id y violations.
    - Fechas ISO-8601 UTC con sufijo Z.
    - Las fases siguen el orden Descubrimiento -> Caracteristicas -> Requisitos -> Modelo.
      La consistencia solo se evalua hacia la derecha (downstream).

    Los endpoints de generacion de codigo estan marcados con x-status: planned;
    el frontend puede integrarse contra ellos mientras el backend los implementa.
  contact:
    name: KOSMO Team
    email: dev@kosmo.app
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Desarrollo local
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/projects/{project_id}/discovery/chat:
    post:
      tags: [discovery]
      summary: Enviar mensaje al chat de Descubrimiento (aplicacion instantanea)
      description: >-
        Procesa el mensaje y aplica el cambio inmediatamente sobre el documento
        cuando corresponde. Si el cambio pertenece a otra fase (giro de negocio),
        responde con una redireccion hacia esa fase.
      operationId: discovery_chat_send
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatRequest"
      responses:
        "200":
          description: Mensaje procesado, modificacion aplicada o redireccion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Error de validacion
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Error al invocar el LLM
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    get:
      tags: [discovery]
      summary: Obtener historial del chat de Descubrimiento
      operationId: discovery_chat_history
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/CursorBefore"
      responses:
        "200":
          description: Historial del chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatHistoryResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/projects/{project_id}/discovery/chat/stream:
    post:
      tags: [discovery]
      summary: Chat de Descubrimiento con streaming SSE
      description: >-
        Genera la respuesta en tiempo real (SSE). Emite eventos token por
        fragmento y, al final, un evento message con la respuesta completa y
        la modificacion aplicada o la redireccion.
      operationId: discovery_chat_stream
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatRequest"
      responses:
        "200":
          description: Stream de eventos SSE
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/ChatStreamEvent"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Error al invocar el LLM
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

  # ══════════════════════════════ FEATURES ══════════════════════════════
...
  /api/v1/features/{feature_id}/chat:
    post:
      tags: [features]
      summary: Enviar mensaje al chat de Caracteristicas (aplicacion instantanea)
      description: >-
        Procesa el mensaje y aplica el cambio inmediatamente sobre la caracteristica
        (titulo, descripcion u origen). Si el cambio pertenece a otra fase,
        responde con una redireccion.
      operationId: feature_chat_send
      parameters:
        - $ref: "#/components/parameters/FeatureId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatRequest"
      responses:
        "200":
          description: Mensaje procesado, modificacion aplicada o redireccion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Error de validacion
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Error al invocar el LLM
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/v1/features/{feature_id}/chat/history:
    get:
      tags: [features]
      summary: Obtener historial del chat de Caracteristicas
      operationId: feature_chat_history
      parameters:
        - $ref: "#/components/parameters/FeatureId"
        - $ref: "#/components/parameters/CursorBefore"
      responses:
        "200":
          description: Historial del chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatHistoryResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/features/{feature_id}/chat/stream:
    post:
      tags: [features]
      summary: Chat de Caracteristicas con streaming SSE
      description: >-
        Genera la respuesta en tiempo real (SSE). Emite eventos token por
        fragmento y, al final, un evento message con la modificacion aplicada
        o la redireccion.
      operationId: feature_chat_stream
      parameters:
        - $ref: "#/components/parameters/FeatureId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatRequest"
      responses:
        "200":
          description: Stream de eventos SSE
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/ChatStreamEvent"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Error al invocar el LLM
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

  # ══════════════════════════════ REQUIREMENTS ══════════════════════════════
...
  /api/v1/features/{feature_id}/requirements:
    get:
      tags: [requirements]
      summary: Obtener requisitos de una caracteristica
      operationId: requirements_get
      parameters:
        - $ref: "#/components/parameters/FeatureId"
        - $ref: "#/components/parameters/ProjectIdQuery"
      responses:
        "200":
          description: Documento de requisitos en Markdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequirementsDocumentResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [requirements]
      summary: Guardar requisitos manualmente
      operationId: requirements_save
      parameters:
        - $ref: "#/components/parameters/FeatureId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveRequirementsRequest"
      responses:
        "200":
          description: Requisitos guardados
          content:
            application/json:
              schema:
                type: object
                properties:
                  feature_id:
                    type: string
                  message:
                    type: string
                    example: ok
                required: [feature_id, message]
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/features/{feature_id}/requirements/chat:
    post:
      tags: [requirements]
      summary: Enviar mensaje al chat de Requisitos (aplicacion instantanea)
      description: >-
        Procesa el mensaje y aplica el cambio inmediatamente sobre el documento
        de requisitos. Si el cambio pertenece a otra fase, responde con una
        redireccion.
      operationId: requirement_chat_send
      parameters:
        - $ref: "#/components/parameters/FeatureId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatRequest"
      responses:
        "200":
          description: Mensaje procesado, modificacion aplicada o redireccion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Error de validacion
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Error al invocar el LLM
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/v1/features/{feature_id}/requirements/chat/history:
    get:
      tags: [requirements]
      summary: Obtener historial del chat de Requisitos
      operationId: requirement_chat_history
      parameters:
        - $ref: "#/components/parameters/FeatureId"
        - $ref: "#/components/parameters/CursorBefore"
      responses:
        "200":
          description: Historial del chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatHistoryResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/features/{feature_id}/requirements/chat/stream:
    post:
      tags: [requirements]
      summary: Chat de Requisitos con streaming SSE
      description: >-
        Genera la respuesta en tiempo real (SSE). Emite eventos token por
        fragmento y, al final, un evento message con la modificacion aplicada
        o la redireccion.
      operationId: requirement_chat_stream
      parameters:
        - $ref: "#/components/parameters/FeatureId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatRequest"
      responses:
        "200":
          description: Stream de eventos SSE
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/ChatStreamEvent"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Error al invocar el LLM
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

  # ══════════════════════════════ MODELO ══════════════════════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT RS256 emitido por /auth/token o /auth/refresh
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
      description: Identificador del proyecto (prj_ + ULID)
    FeatureId:
      name: feature_id
      in: path
      required: true
      schema:
        type: string
      description: Identificador o slug de la caracteristica
    ProjectIdQuery:
      name: project_id
      in: query
      required: true
      schema:
        type: string
      description: Identificador del proyecto
    CursorBefore:
      name: before
      in: query
      required: false
      schema:
        type: string
      description: Cursor ISO-8601 para paginacion hacia atras
  responses:
    NotFound:
      description: Recurso no encontrado
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
  schemas:
    ProblemDetail:
      type: object
      description: Error RFC 7807 (application/problem+json)
      properties:
        type:
          type: string
          description: URN del error (ej. urn:kosmo:projects:not-found)
          example: urn:kosmo:projects:not-found
        title:
          type: string
          description: Mensaje corto
        status:
          type: integer
          description: Codigo HTTP
        detail:
          type: string
          description: Descripcion larga
        instance:
          type: string
          description: Ruta que fallo
        trace_id:
          type: string
          description: Identificador de correlacion (X-Trace-Id)
        violations:
          type: array
          items:
            type: object
            properties:
              loc:
                type: string
                description: Ruta al campo invalido
              msg:
                type: string
              input:
                description: Valor recibido
      required: [type, title, status, detail]

    # ── Auth ──
...
    SendChatRequest:
      type: object
      additionalProperties: false
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 4000
      required: [content]
    ModificationResult:
      type: object
      description: >-
        Resultado de la modificacion aplicada por el chat. Incluye el diff
        aplicado (before/after) para que el frontend ofrezca aceptar (no hacer
        nada, ya esta aplicado) o deshacer via el endpoint /chat/undo con el
        undo_version_id.
      properties:
        applied:
          type: boolean
        modified_section:
          type: string
          nullable: true
        change_description:
          type: string
          nullable: true
        modified_document:
          type: string
          nullable: true
          description: >-
            Markdown completo modificado. En el chat de Caracteristicas contiene
            el nuevo titulo.
        before:
          type: string
          nullable: true
          description: Contenido previo de la seccion o atributo modificado
        after:
          type: string
          nullable: true
          description: Contenido posterior de la seccion o atributo modificado
        undo_version_id:
          type: string
          nullable: true
          description: >-
            Version previa al cambio. Se envia a POST .../chat/undo para
            deshacer la modificacion. Presente solo si applied=true.
        clarification_message:
          type: string
          nullable: true
          description: Presente cuando applied=false (instruccion ambigua)
      required: [applied]
...
    RedirectInfo:
      type: object
      description: Redireccion cuando el cambio pertenece a otra fase
      properties:
        target_phase:
          type: string
          description: Fase destino (descubrimiento, caracteristicas, requisitos, modelo)
        redirect_message:
          type: string
          description: Explicacion de por que el cambio pertenece a esa fase
      required: [target_phase, redirect_message]
    ChatMessageView:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        created_at:
          type: string
          format: date-time
        modification:
          $ref: "#/components/schemas/ModificationResult"
          description: Modificacion aplicada (solo mensajes del asistente)
      required: [id, role, content, created_at]
    ChatResponse:
      type: object
      description: >-
        Respuesta del chat. Contiene el mensaje y, alternativamente, la
        modificacion aplicada o la redireccion a otra fase. Cuando se aplica
        una modificacion, incluye los impactos de consistencia detectados en
        las fases a la derecha (trazabilidad descendente).
      properties:
        message:
          $ref: "#/components/schemas/ChatMessageView"
        modification:
          $ref: "#/components/schemas/ModificationResult"
          description: Presente si el chat aplico un cambio
        redirect:
          $ref: "#/components/schemas/RedirectInfo"
          description: Presente si el cambio pertenece a otra fase
        consistency:
          type: array
          nullable: true
          description: >-
            Impactos de consistencia en las fases a la derecha
            (Descubrimiento -> Caracteristicas -> Requisitos -> Modelo).
            Solo presente cuando applied=true.
          items:
            $ref: "#/components/schemas/ImpactItemView"
      required: [message]
    ChatHistoryResponse:
      type: object
      properties:
        phase:
          type: string
        context:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessageView"
        has_more:
          type: boolean
          default: false
        next_cursor:
          type: string
          nullable: true
      required: [phase, context, messages]
    ChatStreamEvent:
      oneOf:
        - type: object
          description: Fragmento de texto en tiempo real
          properties:
            type:
              type: string
              const: token
            content:
              type: string
          required: [type, content]
        - type: object
          description: Mensaje final con modificacion aplicada o redireccion
          properties:
            type:
              type: string
              const: message
            content:
              type: string
            modification:
              $ref: "#/components/schemas/ModificationResult"
            redirect:
              $ref: "#/components/schemas/RedirectInfo"
          required: [type, content]

    # ── Requirements ──
...
    RequirementsDocumentResponse:
      type: object
      properties:
        document_markdown:
          type: string
          nullable: true
        total:
          type: integer
      required: [document_markdown, total]
    SaveRequirementsRequest:
      type: object
      additionalProperties: false
      properties:
        project_id:
          type: string
        markdown:
          type: string
      required: [project_id, markdown]
...
    ImpactItemView:
      type: object
      properties:
        phase:
          type: string
        artifact_id:
          type: string
        artifact_type:
          type: string
        artifact_label:
          type: string
        section:
          type: string
          default: ""
        rationale:
          type: string
          default: ""
        diff_suggestion:
          type: object
          nullable: true
      required: [phase, artifact_id, artifact_type, artifact_label]
...